combine workflows, remove/rename configs, create new PED method for multicohorts
