+++ configs/defaults/gatk_sv_multisample.toml
@@ -3,6 +3,42 @@ name = 'gatk_sv'
 ref_fasta = 'gs://cpg-common-main/references/hg38/v0/dragen_reference/Homo_sapiens_assembly38_masked.fasta'
 status_reporter = 'metamist'
 
+# switches to deactivate all Metrics workflows. We ran into some resourcing issues when our total
+# cohort size approached ~1600. Response when raised with GATK team:
+#
+# "GatherBatchEvidenceMetrics and the other metrics routines at the end of the main workflows are
+# primarily for development and testing purposes, so you don’t need to run them as a user"
+
+# these metrics workflows are typically running svtest (https://github.com/broadinstitute/gatk-sv/blob/d37e453038e425acba9683da503218fe3d4b1033/src/svtest/setup.py)
+# on each output of the stage, producing output files containing a count/distribution of different variants
+# and sizes. The workflow files we are using as Stages do not always capture these results as top-level outputs,
+# so we have no reliable method to extract these files into permanent storage.
+[resource_overrides.GatherBatchEvidence]
+run_matrix_qc = false
+run_module_metrics = false
+
+[resource_overrides.ClusterBatch]
+run_module_metrics = false
+
+[resource_overrides.GenerateBatchMetrics]
+run_module_metrics = false
+
+[resource_overrides.FilterBatch]
+run_module_metrics = false
+
+[resource_overrides.GenotypeBatch]
+run_module_metrics = false
+
+[resource_overrides.MakeCohortVcf]
+run_module_metrics = false
+
+[resource_overrides.FilterGenotypes]
+run_qc = false
+
+# if you have to do it, bump the resources
+#[resource_overrides.FilterGenotypes.runtime_override_plot_qc_per_family]
+#mem_gb = 32
+
 [references.gatk_sv]
 # This model is trained on the All Of Us dataset. If this is used in any published work,
 # credit must be given in accordance with the AoU publication policy.
@@ -53,19 +89,3 @@ username = 'seqr'
 # Load ElasticSearch password from a secret, unless SEQR_ES_PASSWORD is set
 password_secret_id = 'seqr-es-password'
 password_project_id = 'seqr-308602'
-
-# switch off metrics workflows
-[resource_overrides.MakeCohortVcf]
-# this is passed to CleanVcf
-run_module_metrics = false
-
-# if you have to do it, bump the resources
-#[resource_overrides.MakeCohortVcf.runtime_override_plot_qc_per_family]
-#mem_gb = 32
-
-[resource_overrides.FilterGenotypes]
-run_qc = false
-
-# if you have to do it, bump the resources
-#[resource_overrides.FilterGenotypes.runtime_override_plot_qc_per_family]
-#mem_gb = 32
